+++ protocols/jaeger_unmarshaler.go
@@ -1,10 +1,10 @@
-// Copyright 2020 The OpenTelemetry Authors
+// Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//      http://www.apache.org/licenses/LICENSE-2.0
+//       http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kafkareceiver
+package protocols
 
 import (
 	"bytes"
@@ -24,23 +24,29 @@ import (
 	jaegertranslator "go.opentelemetry.io/collector/translator/trace/jaeger"
 )
 
+// JaegerProtobuf protocol type.
+const JaegerProtobuf Type = "jaeger/protobuf"
+
 type jaegerProtoSpanUnmarshaler struct {
 }
 
 var _ TracesUnmarshaler = (*jaegerProtoSpanUnmarshaler)(nil)
 
 func (j jaegerProtoSpanUnmarshaler) Unmarshal(bytes []byte) (pdata.Traces, error) {
-	span := &jaegerproto.Span{}
-	err := span.Unmarshal(bytes)
+	batch := jaegerproto.Batch{}
+	err := batch.Unmarshal(bytes)
 	if err != nil {
 		return pdata.NewTraces(), err
 	}
-	return jaegerSpanToTraces(span), nil
+	return jaegertranslator.ProtoBatchToInternalTraces(batch), nil
 }
 
-func (j jaegerProtoSpanUnmarshaler) Encoding() string {
-	return "jaeger_proto"
+func (j jaegerProtoSpanUnmarshaler) Type() Type {
+	return JaegerProtobuf
 }
+
+// JaegerJSON protocol type.
+const JaegerJSON Type = "jaeger/json"
 
 type jaegerJSONSpanUnmarshaler struct {
 }
@@ -48,22 +54,14 @@ type jaegerJSONSpanUnmarshaler struct {
 var _ TracesUnmarshaler = (*jaegerJSONSpanUnmarshaler)(nil)
 
 func (j jaegerJSONSpanUnmarshaler) Unmarshal(data []byte) (pdata.Traces, error) {
-	span := &jaegerproto.Span{}
-	err := jsonpb.Unmarshal(bytes.NewReader(data), span)
+	batch := jaegerproto.Batch{}
+	err := jsonpb.Unmarshal(bytes.NewReader(data), &batch)
 	if err != nil {
 		return pdata.NewTraces(), err
 	}
-	return jaegerSpanToTraces(span), nil
+	return jaegertranslator.ProtoBatchToInternalTraces(batch), nil
 }
 
-func (j jaegerJSONSpanUnmarshaler) Encoding() string {
-	return "jaeger_json"
-}
-
-func jaegerSpanToTraces(span *jaegerproto.Span) pdata.Traces {
-	batch := jaegerproto.Batch{
-		Spans:   []*jaegerproto.Span{span},
-		Process: span.Process,
-	}
-	return jaegertranslator.ProtoBatchToInternalTraces(batch)
+func (j jaegerJSONSpanUnmarshaler) Type() Type {
+	return JaegerJSON
 }