+++ protocols/otlp_unmarshaler_test.go
@@ -1,10 +1,10 @@
-// Copyright 2020 The OpenTelemetry Authors
+// Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//      http://www.apache.org/licenses/LICENSE-2.0
+//       http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kafkareceiver
+package protocols
 
 import (
 	"testing"
@@ -36,7 +36,7 @@ func TestUnmarshalOTLPTraces(t *testing.T) {
 	require.NoError(t, err)
 
 	assert.Equal(t, td, got)
-	assert.Equal(t, "otlp_proto", p.Encoding())
+	assert.Equal(t, Type("otlp/protobuf"), p.Type())
 }
 
 func TestUnmarshalOTLPTraces_error(t *testing.T) {
@@ -56,7 +56,7 @@ func TestUnmarshalOTLPLogs(t *testing.T) {
 	require.NoError(t, err)
 
 	assert.Equal(t, ld, got)
-	assert.Equal(t, "otlp_proto", p.Encoding())
+	assert.Equal(t, Type("otlp/protobuf"), p.Type())
 }
 
 func TestUnmarshalOTLPLogs_error(t *testing.T) {