+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Rename `marshall` to `marshal` in all the occurrences (#2977)
 - Remove `componenterror.ErrAlreadyStarted` and `componenterror.ErrAlreadyStopped`, components should not protect against this, Service will start/stop once.
 - Rename `ApplicationStartInfo` to `BuildInfo`
+- Change `encodings` option in [Kafka receiver](receiver/kafkareceiver) to use new [protocol](protocols/README.md) types. (#3044)
 
 ## 💡 Enhancements 💡